Move ask assistant action to its own file
